requirements.txt
# Core dependencies for Security Compliance Automation Agent
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# JWT and cryptographic operations
PyJWT[crypto]==2.8.0
cryptography>=41.0.7

# Database and ORM
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# HTTP client and async support
aiohttp==3.9.1
httpx==0.25.2

# Monitoring and metrics
prometheus-client==0.19.0

# Event streaming (Kafka)
aiokafka==0.10.0

# Logging and configuration
structlog==23.2.0
python-json-logger==2.0.7
python-dotenv==1.0.0

# Security and validation
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx[dev]==0.25.2

# Scheduling
apscheduler==3.10.4

# XML parsing for OpenSCAP results
lxml==4.9.4

# Data validation and serialization
marshmallow==3.20.2

# High-performance event loop (optional)
uvloop==0.19.0